Name Convention Conversion
You can now switch from Latin to common or common to Latin and it will keep the selected choice as well as change the name from one language to the other. I had this working at one point - adding the multiselect box in the header broke it. I've refactored it so that it will work in the future irrespective of other changes / additions.