Sliding Image By Species Box
Added a sliding image beside the species box that will change based on what species is being hovered over as well as move up or down with the cursor and disappear whenever the user moves away from the selection box. Very similar to the one in Phase 1, but in Vue.js instead of jQuery.